Rename OAuth2SecurityRequirementsFilter to AssignOAuth2SecurityOperationFilter and optimize it
