+++ VirtoCommerce.Platform.Web/Swagger/AssignOAuth2SecurityOperationFilter.cs
@@ -4,9 +4,7 @@ using System.Linq;
 using System.Web.Http;
 using System.Web.Http.Description;
 using Swashbuckle.Swagger;
-
-using MvcAllowAnonymousAttribute = System.Web.Mvc.AllowAnonymousAttribute;
-using WebApiAllowAnonymousAttribute = System.Web.Http.AllowAnonymousAttribute;
+using VirtoCommerce.Platform.Core.Web.Security;
 
 namespace VirtoCommerce.Platform.Web.Swagger
 {
@@ -15,26 +13,33 @@ namespace VirtoCommerce.Platform.Web.Swagger
         [CLSCompliant(false)]
         public void Apply(Operation operation, SchemaRegistry schemaRegistry, ApiDescription apiDescription)
         {
-            var actionFilters = apiDescription.ActionDescriptor.GetFilterPipeline();
-            bool allowAnonymous = actionFilters.Select(x => x.Instance).OfType<OverrideAuthorizationAttribute>().Any();
-            if (allowAnonymous)
+            if (apiDescription.GetControllerAndActionAttributes<AllowAnonymousAttribute>().Any())
+            {
+                operation.security = null;
+                return;
+            }
+
+            var permissionAttributes = apiDescription.ActionDescriptor
+                .GetFilterPipeline().Select(x => x.Instance).OfType<CheckPermissionAttribute>();
+
+            var permissions = permissionAttributes.SelectMany(x => x.Permissions).ToList();
+            if (permissions.Count == 0)
             {
                 return;
             }
-            else
+
+            var securityRequirements = new Dictionary<string, IEnumerable<string>>
             {
-                if (operation.security == null)
-                {
-                    operation.security = new List<IDictionary<string, IEnumerable<string>>>();
-                }
+                {"OAuth2", permissions},
+                {"apiKey", permissions}
+            };
 
-                var oAuthRequirements = new Dictionary<string, IEnumerable<string>>
-                {
-                    { "oauth2", new [] { OwinConfig.PublicClientId } }
-                };
-
-                operation.security.Add(oAuthRequirements);
+            if (operation.security == null)
+            {
+                operation.security = new List<IDictionary<string, IEnumerable<string>>>(1);
             }
+
+            operation.security.Add(securityRequirements);
         }
     }
 }
